rename plot_utils.py -> mpl_widget.py Signed-off-by: Chipmuenk <mail@chipmuenk.de>
