+++ pyfda/plot_widgets/plot_hf.py
@@ -20,7 +20,7 @@ from matplotlib import rcParams
 
 import pyfda.filterbroker as fb
 from pyfda.pyfda_rc import params
-from pyfda.plot_widgets.plot_utils import MplWidget
+from pyfda.plot_widgets.mpl_widget import MplWidget
 from pyfda.pyfda_lib import calc_Hcomplex
 
 class PlotHf(QWidget):
